+++ src/datasources/SQLSRVDataSource.php
@@ -250,7 +250,7 @@ class SQLSRVDataSource extends DataSource
                 $value = implode(",", $value);
                 $query = str_replace($key, $value, $query);
             } else {
-                $query = str_replace($key, $this->escape($value), $query);
+                $query = str_replace($key, (string)$this->escape($value), $query);
             }
         }
         return $query;